ircsprintf.c: fixed compile warning on gcc281 send.c: fixed compile warning on gcc281 include/setup.h: shouldn't have been there, gets generated by configure

+++ src/ircsprintf.c
@@ -73,7 +73,7 @@ va_dcl
 	    rp = inp[++i];                  /* get the next parameter */
 	    break;
 	  case 'c':
-	    *wp++ = (char)rp;
+	    *wp++ = (char) ((int)rp);
 	    rp = inp[++i];
 	    break;
 	  case 'd':
